+++ packages/ui/src/lib/index.ts
@@ -1,4 +1,3 @@
 export * from './classNames'
 export * from './use-get-dom-info-hooks'
-export * from './use-get-root-header-info'
 export * from './has-mixpanel'
