+++ packages/ui/src/lib/use-get-dom-info-hooks.ts
@@ -2,20 +2,20 @@
 import { useEffect, useRef, useState } from "react"
 
 // 获取bio中的Root Header的高度
-// export function useGetRootHeaderInfo() {
-//   const [rootHeaderHeight, setRootHeaderHeight] = useState(0)
-//   const rootHeaderRef = useRef<Element | null>(null)
+export function useGetRootHeaderInfo() {
+  const [rootHeaderHeight, setRootHeaderHeight] = useState(0)
+  const rootHeaderRef = useRef<Element | null>(null)
 
-//   useEffect(() => {
-//     const rootHeaderEl = document.querySelector('#root-header')
-//     const rootHeaderElHeight = rootHeaderEl?.getBoundingClientRect().height || 0
+  useEffect(() => {
+    const rootHeaderEl = document.querySelector('#root-header')
+    const rootHeaderElHeight = rootHeaderEl?.getBoundingClientRect().height || 0
 
-//     setRootHeaderHeight(rootHeaderElHeight)
-//     rootHeaderRef.current = rootHeaderEl
-//   }, [])
+    setRootHeaderHeight(rootHeaderElHeight)
+    rootHeaderRef.current = rootHeaderEl
+  }, [])
 
-//   return { rootHeaderRef, rootHeaderHeight }
-// }
+  return { rootHeaderRef, rootHeaderHeight }
+}
 
 // 获取pdp页中的吸顶tabs的高度
 export function useGetProductTabsInfo() {
